+++ src/gwt/src/org/rstudio/studio/client/application/DesktopFrame.java
@@ -175,11 +175,10 @@ public interface DesktopFrame extends JavaScriptPassthrough
    void setMousewheelZoomEnabled(boolean enabled);
    void setMousewheelZoomDebounce(double debounceMs);
    
-   void getEnableAccessibility(CommandWithArg<Boolean> callback);
    void setEnableAccessibility(boolean enable);
    void setDisableRendererAccessibility(boolean disable);
-   void getEnableSplashScreen(CommandWithArg<Boolean> callback);
    void setEnableSplashScreen(boolean enable);
+   void setShowWhatsNew(boolean enable);
 
    void setAutohideMenubar(boolean enable);